pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deprecated>=1.3.1",
    "dotenv>=0.9.9",
    "fastapi>=0.121.1",
    "google-adk>=1.18.0",
    "google-genai>=1.49.0",
    "httpx>=0.28.1",
    "pydantic[email]>=2.12.4",
    "pyjwt>=2.10.1",
    "pytest>=9.0.0",
    "pytest-order>=1.3.0",
    "python-dotenv==1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
